[Refactor] refactor patch module
