+++ vllm_ascend/patch/platform/patch_config.py
