+++ vllm_ascend/patch/platform/patch_distributed.py
