+++ vllm_ascend/patch/platform/patch_mamba_config.py
