+++ vllm_ascend/patch/platform/patch_multiproc_executor.py
