+++ vllm_ascend/patch/worker/patch_logits.py
