+++ vllm_ascend/patch/worker/patch_minicpm.py
