+++ vllm_ascend/patch/worker/patch_multimodal_merge.py
