+++ vllm_ascend/patch/worker/patch_roberta.py
