+++ vllm_ascend/patch/worker/patch_weight_loader.py
